Add a centralized dependent_false type
The main use case is to static_assert on it at the end of a
if constexpr / else if constexpr chain.

See P2593 for a discussion about why this is pretty much the
only "allowed" way of doing so, short of running into IFNDR.

I'm actually adding two versions: one for TTP and one for NTTP,
as Qt code uses both.

Apply it in QFlatMap.

Change-Id: Iaff97e350784683d0c3994020b1352d5188931d6
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>